+++ packages/sdk-statement/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./codec"
 export * from "./statement-sdk"
 export * from "./signer"
-export type { SubmitResult } from "./types"
+export type { SubmitResult, TopicFilter, StatementEvent } from "./types"
 export { stringToTopic } from "./utils"