+++ packages/sdk-statement/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### Changed
+
+- Change API to new subscription-based statement store
+  - `createStatementSdk` takes an endpoint URL rather than a `req` function.
+  - Update `Statement` to new spec: replaces `priority` for `expiry`.
+  - Removes `dump()`
+  - `getStatements` requires the `topics` parameter
+  - Adds `getBroadcasts` and `getPosted` for filtering based on the `dest` parameter
+  - Adds `subscribeStatements(topics)` for subscribing
+
 ## 0.4.1 2026-03-16
 
 ### Fixed